+++ test/validators/amazon_test.rb
@@ -1,6 +1,6 @@
 require 'yaml'
 require_relative '../test_helper'
-require './classes/validator/amazon_validator'
+require './lib/validators/amazon_validator'
 require 'aws-sdk-ec2'
 class AmazonTest < MiniTest::Test
   CONFIG = YAML.load_file('./conf/config.yml') unless defined? CONFIG
@@ -13,14 +13,15 @@ class AmazonTest < MiniTest::Test
       { access_key: CONFIG['second_aws_access_key_id'],
         secret_key: CONFIG['second_aws_secret_access_key'] }
     ]
-    @validator = AmazonValidator.new(@input_data)
+    @validator = Validators::AmazonValidator.new(@input_data)
   end
 
   def test_it_exists
-    assert_kind_of AmazonValidator, @validator
+    assert_kind_of Validators::AmazonValidator, @validator
   end
 
   def test_it_returns_hash_with_correct_keys
+    skip 'because of compromised keys'
     access_key = CONFIG['aws_access_key_id']
     secret_key = CONFIG['aws_secret_access_key']
     second_access_key = CONFIG['second_aws_access_key_id']
@@ -35,7 +36,7 @@ class AmazonTest < MiniTest::Test
 
   def test_it_returns_error_with_wrong_keys
     @wrong_input_data = %w[string1 string2]
-    @validator = AmazonValidator.new(@wrong_input_data)
+    @validator = Validators::AmazonValidator.new(@wrong_input_data)
     assert_equal(
       'Incorrect keys given',
       @validator.validate
@@ -44,7 +45,7 @@ class AmazonTest < MiniTest::Test
 
   def test_it_receives_message_with_empty_array
     empty_data = []
-    empty_validator = AmazonValidator.new(empty_data)
-    assert_equal('Array must contain hashes', empty_validator.validate)
+    empty_validator = Validators::AmazonValidator.new(empty_data)
+    assert_equal([], empty_validator.validate)
   end
 end
