+++ test/validators/telegram_test.rb
@@ -1,7 +1,7 @@
 require 'yaml'
 require 'telegram/bot'
 require_relative '../test_helper'
-require './classes/validator/telegram_validator'
+require './lib/validators/telegram_validator'
 
 class TelegramTest < Minitest::Test
   CONFIG = YAML.load_file('./conf/config.yml') unless defined? CONFIG
@@ -14,17 +14,17 @@ class TelegramTest < Minitest::Test
       { token: @wrong_token },
       { token: @second_valid_token }
     ]
-    @validator = TelegramValidator.new(@input_data)
+    @validator = Validators::TelegramValidator.new(@input_data)
   end
 
   def test_it_exists
-    assert_kind_of TelegramValidator, @validator
+    assert_kind_of Validators::TelegramValidator, @validator
   end
 
   def test_it_receives_message_with_empty_array
     empty_data = []
-    empty_validator = TelegramValidator.new(empty_data)
-    assert_equal('Array must contain hashes', empty_validator.validate)
+    empty_validator = Validators::TelegramValidator.new(empty_data)
+    assert_equal([], empty_validator.validate)
   end
 
   def test_it_returns_true_with_valid_keys
